Add disassembly to bundle and package builders
Also handles fact exceptions to avoid an extra break when debugging tests.

+++ test/src/WixTestTools/WixTestBase.cs
@@ -323,21 +323,24 @@ namespace WixTest
             {
                 foreach (FileSystemInfo artifact in this.TestArtifacts)
                 {
-                    try
+                    if (artifact.Exists)
                     {
-                        DirectoryInfo dir = artifact as DirectoryInfo;
-                        if (null != dir)
+                        try
                         {
-                            dir.Delete(true);
+                            DirectoryInfo dir = artifact as DirectoryInfo;
+                            if (null != dir)
+                            {
+                                dir.Delete(true);
+                            }
+                            else
+                            {
+                                artifact.Delete();
+                            }
                         }
-                        else
+                        catch
                         {
-                            artifact.Delete();
+                            Debug.WriteLine(String.Format("Failed to delete '{0}'.", artifact.FullName));
                         }
-                    }
-                    catch
-                    {
-                        Debug.WriteLine(String.Format("Failed to delete '{0}'.", artifact.FullName));
                     }
                 }
             }
